ESP player distance, etc
Updated PlayerInfo
Updated Drawing
Updated Font
Added ESP player distance

+++ Engine/Drawing/Drawing.h
@@ -8,17 +8,17 @@ namespace Engine
 	{
 	public:
 		//Fill area
-		void FillArea ( int x, int y, int w, int h, BYTE r, BYTE g, BYTE b, BYTE a, BYTE style );
+		void FillArea ( float x, float y, float w, float h, BYTE r, BYTE g, BYTE b, BYTE a, BYTE style );
 		//Drawing box (style: 1 - GL_QUADS, 2 - GL_LINE_LOOP)
-		void DrawBox ( int x, int y, int w, int h, int linewidth, BYTE r, BYTE g, BYTE b, BYTE a , BYTE style );
+		void DrawBox ( float x, float y, float w, float h, float linewidth, BYTE r, BYTE g, BYTE b, BYTE a, BYTE style );
 		//Drawing box with outline (style: 1 - GL_QUADS, 2 - GL_LINE_LOOP)
-		void DrawShadowBox ( int x, int y, int w, int h, int linewidth, BYTE r, BYTE g, BYTE b, BYTE a, BYTE style );
+		void DrawShadowBox ( float x, float y, float w, float h, float linewidth, BYTE r, BYTE g, BYTE b, BYTE a, BYTE style );
 		//Drawing 3d box on player (style: 1 - GL_LINES, 2 - GL_LINE_LOOP)
 		void Draw3DBox ( struct cl_entity_s *Entity, int Index, float lw, BYTE r, BYTE g, BYTE b, BYTE style );
 		//Drawing circle
 		void DrawCircle ( float x, float y, float rad, int amountSegments, float linewidth, BYTE r, BYTE g, BYTE b, BYTE a );
 		//Drawing line
-		__inline void DrawLine ( int x1, int y1, int x2, int y2, float linewidth, BYTE r, BYTE g, BYTE b, BYTE style );
+		__inline void DrawLine ( float x1, float y1, float x2, float y2, float linewidth, BYTE r, BYTE g, BYTE b, BYTE style );
 		//Drawing vector line
 		__inline void DrawVectorLine ( float *flSrc, float *flDestination, float linewidth, BYTE r, BYTE g, BYTE b, BYTE style );
 	};
